added role enum, Logger implemented for all request to log user-agent
