+++ src/main/java/com/amazkart/dto/Role.java
@@ -1,0 +1,11 @@
+package com.amazkart.dto;
+
+public enum Role {
+    ROLE_GUEST,
+    ROLE_USER,
+    ROLE_ADMIN,
+    ROLE_SUPER,
+    ROLE_SELLER,
+    ROLE_BUYER,
+    ROLE_SUPPORT;
+}
