+++ src/main/java/com/amazkart/jwtcfg/JwtRequestFilter.java
@@ -10,6 +10,8 @@ import org.springframework.security.web.authentication.WebAuthenticationDetailsS
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
 
+import com.amazkart.utility.UserAgentLogger;
+
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -27,6 +29,9 @@ public class JwtRequestFilter extends OncePerRequestFilter {
 	@Override
 	protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain chain)
 			throws ServletException, IOException {
+
+		UserAgentLogger.logUserAgent(request.getHeader("User-Agent"));
+
 		final String authorizationHeader = request.getHeader("Authorization");
 
 		String username = null;
